+++ backend/src/main/java/com/paradise/event_ticket_system/checkout/OrderResponse.java
@@ -5,7 +5,7 @@ import com.paradise.event_ticket_system.order.OrderStatus;
 public record OrderResponse(
         String orderNumber,
         OrderStatus status,
-        Long eventId,
+        Integer eventId,
         String guestName,
         String guestEmail,
         OrderSummaryResponse summary
